+++ scripts/train_all_models.py
@@ -13,7 +13,7 @@ from sklearn.preprocessing import StandardScaler, RobustScaler
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
-from libML.data_load_and_label_for_training import load_emg_data
+from libML. import load_emg_data
 from libML.models import choose_model
 from libML.evaluation import compute_scores, plot_cv_scores, plot_labels, evaluate_overall_system_performance_detailed
 from libML.export import save_best_params, save_model
@@ -81,19 +81,19 @@ MAP_DOF_NAME_TO_ID = {
     }
 
 def get_features(data_dict):
-    # Get windows 
+    # Get windows
     X = data_dict["X"]
     y = data_dict["y"]
     windowed_df = segment_aux_windows_new(X, y)
-    
+
     #breakpoint()
     # Apply preprocessing steps (filtering)
     preproc_df = notch_filter(windowed_df)
-    preproc_df = passband_filter(preproc_df) 
+    preproc_df = passband_filter(preproc_df)
 
     # Get features from windows
     # features_df contains all 6 channels with each feature --> 6*21 = 126 columns, named for example 0_AR4 --> channel id + _ + feature name
-    # + 2 columns for label and window index so we have 128 columns    
+    # + 2 columns for label and window index so we have 128 columns
     features_list = []
     for idx in range(len(preproc_df)):
         window_features = {}
@@ -104,7 +104,7 @@ def get_features(data_dict):
             #print(idx, ch)
             #print(signal_window)
             ch_features = extract_window_features(signal_window, fs=1000)
-            
+
             # Prefix with channel name
             for feat_name, feat_val in ch_features.items():
                 window_features[f"{ch}_{feat_name}"] = feat_val
@@ -121,7 +121,7 @@ def get_features(data_dict):
         window_features['dof_8'] = preproc_df.loc[idx, 'dof_8_label']
         #window_features['dof_9'] = preproc_df.loc[idx, 'dof_9_label']
         #window_features['dof_10'] = preproc_df.loc[idx, 'dof_10_label']
-        
+
         features_list.append(window_features)
 
     features_df = pd.DataFrame(features_list)
@@ -137,14 +137,14 @@ def train_dof_model(dof, X_train, y_train, X_test, y_test):
 
     y_dof_train = y_train[MAP_DOF_NAME_TO_ID[dof]]
     y_dof_test = y_test[MAP_DOF_NAME_TO_ID[dof]]
-    
+
     if len(np.unique(y_dof_train)) == 1 or len(np.unique(y_dof_test)) == 1:
         print("Single label -> no classification")
         return dof, None, None, None, timing_data
-    
+
     model = choose_model(MODEL_TYPE, ALL_HYPERPARAMS, RANDOM_STATE)
     best_params = None
-    
+
     if HYPERPARAMETER_SEARCH:
         search_start = time.time()
         # Your existing hyperparameter search code with n_jobs=-1
@@ -163,43 +163,43 @@ def train_dof_model(dof, X_train, y_train, X_test, y_test):
         best_params = final_model.get_params()
         timing_data['training_time'] = training_time
         timing_data['best_params'] = best_params
-    
+
     total_time = time.time() - start_time
     timing_data['total_time'] = total_time
 
     # Evaluate
     y_pred = final_model.predict(X_test)
     test_scores = compute_scores(y_dof_test, y_pred)
-    
+
     return dof, final_model, test_scores, best_params, timing_data
 
 def save_timing_and_params(timing_data, best_params, model_type, dof, output_dir):
     """Save timing information and best parameters to files"""
     os.makedirs(output_dir, exist_ok=True)
     timestamp = time.strftime('%Y%m%d_%H%M%S')
-    
+
     # Save timing info as CSV
     timing_filename = f"timing_{model_type}_{dof}_{timestamp}.csv"
     timing_filepath = os.path.join(output_dir, timing_filename)
-    
+
     df = pd.DataFrame([timing_data])
     df.to_csv(timing_filepath, index=False)
     print(f"Timing info saved to {timing_filepath}")
-    
+
     # Save best parameters as JSON
     params_filename = f"best_params_{model_type}_{dof}_{timestamp}.json"
     params_filepath = os.path.join(output_dir, params_filename)
-    
+
     with open(params_filepath, 'w') as f:
         json.dump(best_params, f, indent=2)
     print(f"Best parameters saved to {params_filepath}")
-    
+
     # Also save combined info for easy analysis
     combined_data = timing_data.copy()
     combined_data['best_params'] = best_params
     combined_filename = f"combined_{model_type}_{dof}_{timestamp}.json"
     combined_filepath = os.path.join(output_dir, combined_filename)
-    
+
     with open(combined_filepath, 'w') as f:
         json.dump(combined_data, f, indent=2, default=str)  # default=str handles non-serializable objects
 
@@ -214,15 +214,15 @@ def scale_features(processed_data, method='standard', save_path=None, return_dat
     else:
         all_data = processed_data
         was_list = False
-    
+
     # Identify columns
     label_cols = [col for col in all_data.columns if col.startswith('dof_')]
     meta_cols = [col for col in all_data.columns if col in ['window_index', 'label', 'subject_id', 'session_id']]
     feature_cols = [col for col in all_data.columns if col not in label_cols + meta_cols]
-    
+
     X = all_data[feature_cols].values
     feature_names = feature_cols  # Keep track of feature names
-    
+
     # Choose and fit scaler
     if method == 'standard':
         scaler = StandardScaler()
@@ -233,16 +233,16 @@ def scale_features(processed_data, method='standard', save_path=None, return_dat
         scaler = MinMaxScaler()
     else:
         raise ValueError(f"Unknown scaling method: {method}")
-    
+
     X_scaled = scaler.fit_transform(X)
-    
+
     # Save scaler
     if save_path:
         os.makedirs(save_path, exist_ok=True)
         scaler_filename = f"scaler_{method}.pkl"
         scaler_filepath = os.path.join(save_path, scaler_filename)
         joblib.dump(scaler, scaler_filepath)
-        
+
         # Also save feature names for reference
         feature_info = {
             'feature_names': feature_cols,
@@ -253,13 +253,13 @@ def scale_features(processed_data, method='standard', save_path=None, return_dat
         feature_info_path = os.path.join(save_path, f"feature_info_{method}.pkl")
         joblib.dump(feature_info, feature_info_path)
         print(f"Scaler and feature info saved to {save_path}")
-    
+
     # Return appropriate format
     if return_dataframe:
         # Create new DataFrame with scaled features
         scaled_df = all_data.copy()
         scaled_df[feature_cols] = X_scaled
-        
+
         if was_list:
             # Return list of DataFrames if input was list
             result = []
@@ -287,9 +287,9 @@ def main():
         TODO
         1) loop for subjects and for sessions
         2) load emg + triggers
-        3) empty all_sessions_data. For each emg file : 
+        3) empty all_sessions_data. For each emg file :
         (3.1) resample X_raw (emg timeseries) using the emg timestamps -> X_resampled)
-        3.2) trigger vector y_triggers: same length as X_resampled, iterate throught timestamps and take the last triggers 
+        3.2) trigger vector y_triggers: same length as X_resampled, iterate throught timestamps and take the last triggers
         3.3) map trigger to labels : y (in a dict for 8 DoFs) (+ add previous state of the arm pronation/neutral/supination)
         3.4) Window (only take windows with same labels) + filter (causal filters : bandpass + Notch) + feature extraction -> X_features (2D array), y_labels (dict of 1D arrays)
         3.5) Scale features, save scalers
@@ -321,7 +321,7 @@ def main():
 
         # Save preprocessed data
         #save_features_to_disk(processed_data_list, PROCESSED_DATA_PATH)
-    
+
     # TODO: Scale features and save scalers
     # 4. Scale features
     print("Scaling features...")
@@ -334,14 +334,14 @@ def main():
 
     # 5. Train-test-split (using scaled data)
     all_data = pd.concat(scaled_data_list, ignore_index=True)
-    
+
     # Separate features from labels
     label_cols = [col for col in all_data.columns if col.startswith('dof_')]
     feature_cols = [col for col in all_data.columns if col not in label_cols + ['window_index', 'label']]
-    
+
     X = all_data[feature_cols].values
     y = {dof: all_data[dof].values for dof in label_cols}
-    
+
     # Simple random split
     # X_train, X_test = train_test_split(X, test_size=TEST_SIZE, random_state=RANDOM_STATE)
     # y_train = {dof: train_test_split(y[dof], test_size=TEST_SIZE, random_state=RANDOM_STATE)[0] for dof in label_cols}
@@ -383,7 +383,7 @@ def main():
             trained_models[dof] = model
             all_timing_data[dof] = timing_data
             all_best_params[dof] = best_params
-            
+
             # Save individual model results
             save_model(model, MODEL_TYPE, dof, MODEL_OUTPUT_DIR, scaler=None)
             save_timing_and_params(timing_data, best_params, MODEL_TYPE, dof, TIMING_OUTPUT_DIR)
@@ -396,16 +396,16 @@ def main():
         if dof in trained_models:
             model = trained_models[dof]
             y_test_dof = y_test[MAP_DOF_NAME_TO_ID[dof]]
-            
+
             start_time = time.time()
             y_pred_test = model.predict(X_test)
             inference_time = time.time() - start_time
             # Probability of prediciting each class
-            y_pred_proba_test = getattr(model, "predict_proba", lambda x: None)(X_test) 
-            
+            y_pred_proba_test = getattr(model, "predict_proba", lambda x: None)(X_test)
+
             test_scores = compute_scores(y_test_dof, y_pred_test)
             test_time = time.time() - start_time
-            
+
             all_test_scores[dof] = {}
             all_test_scores[dof]["Accuracy"] = test_scores[0]
             all_test_scores[dof]["Precision"] = test_scores[1]
@@ -468,4 +468,3 @@ def main():
 
 if __name__ == "__main__":
     main()
-
